+++ services/tictactoe/api/src/Match.py
@@ -1,16 +1,19 @@
-from    .Player  import   PlayerC
-from    .Move    import   MoveC
+import                      asyncio
+from    .Player  import     PlayerC
+from    .Move    import     MoveC
 
 PENDING     = 0
 DRAW        = 1
 WIN         = 3
 
-class MatchC():
+Matches     = { "PENDING":{}, "MATCHING":{}, "PLAYING":{} }
+Users       = { "PENDING"}
 
-    def __init__( self ):
-        self.__id        = ""
-        self.__player_x  = PlayerC("/home/ychaaibi/Desktop/ft_transcendence/services/tictactoe/api/bin/blue") #player_x blue
-        self.__player_o  = PlayerC("/home/ychaaibi/Desktop/ft_transcendence/services/tictactoe/api/bin/red") #player_o red
+class MatchC():
+    def __init__( self,  ):
+        self.__id        = "" #generate using database
+        self.__room_name = "" #using player
+        self.__state     = "PENDING"  
         self.__moves     = [ ] #moves
 
     def __str__( self ):
@@ -18,6 +21,25 @@ class MatchC():
               + "moves      : " + str(self.__moves)
             #   + "player_x   : " + self.__player_x    + "\n"  \ 
             #   + "player_o   : " + self.__player_o    + "\n"  \
+
+    async def wait_match( self ):
+        await asyncio.sleep( 120 )
+        #incst
+        #incst
+        #incst
+        #incst
+        #incst
+        #incst
+        # Can have mutexes
+        Matches[ self.__state ].pop( self.__id )
+        #end
+
+    async def wait_player( self ):
+        await asyncio.sleep( 15 )
+
+        # Can have mutexes
+        # Abort the player and the other one win
+        # end
 
     def simulate( self, move_s, type_player ):
         move    = MoveC( type_player,
